third_party/blink/web_tests/fast/scroll-behavior/main-frame-scroll-in-quirks-mode-expected.txt

Test that calling scroll methods on the body element scrolls the viewport in quirks mode

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

Test that "scroll" on the document element does not scroll
PASS document.documentElement.scrollTop is 0
PASS document.documentElement.scrollLeft is 0
PASS document.body.scrollTop is 0
PASS document.body.scrollLeft is 0


Test that "scrollTo" on the document element does not scroll
PASS document.documentElement.scrollTop is 0
PASS document.documentElement.scrollLeft is 0
PASS document.body.scrollTop is 0
PASS document.body.scrollLeft is 0


Test that "scrollBy" on the document element does not scroll
PASS document.documentElement.scrollTop is 0
PASS document.documentElement.scrollLeft is 0
PASS document.body.scrollTop is 0
PASS document.body.scrollLeft is 0


Test that "scroll" on the body element scrolls
PASS document.documentElement.scrollTop is 0
PASS document.documentElement.scrollLeft is 0
PASS document.body.scrollTop is 100
PASS document.body.scrollLeft is 50


Test that "scrollTo" on the body element scrolls
PASS document.documentElement.scrollTop is 0
PASS document.documentElement.scrollLeft is 0
PASS document.body.scrollTop is 200
PASS document.body.scrollLeft is 450


Test that "scrollBy" on the body element scrolls
PASS document.documentElement.scrollTop is 0
PASS document.documentElement.scrollLeft is 0
PASS document.body.scrollTop is 300
PASS document.body.scrollLeft is 750
